wip: db tests
